move actions. move protected routes logic
