+++ examples/with-supabase/app/(auth-pages)/signup/page.tsx
@@ -1,44 +1,11 @@
-import { createClient } from "@/utils/supabase/server";
-import { headers } from "next/headers";
-import Link from "next/link";
+import { signUpAction } from "@/app/actions";
+import { FormMessage, Message } from "@/components/form-message";
 import { SubmitButton } from "@/components/submit-button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { FormMessage, Message } from "@/components/form-message";
-import { encodedRedirect } from "@/utils/utils";
+import Link from "next/link";
 
 export default function Signup({ searchParams }: { searchParams: Message }) {
-  const signUp = async (formData: FormData) => {
-    "use server";
-    const email = formData.get("email")?.toString();
-    const password = formData.get("password")?.toString();
-    const supabase = createClient();
-    const origin = headers().get("origin");
-
-    if (!email || !password) {
-      return { error: "Email and password are required" };
-    }
-
-    const { error } = await supabase.auth.signUp({
-      email,
-      password,
-      options: {
-        emailRedirectTo: `${origin}/auth/callback`,
-      },
-    });
-
-    if (error) {
-      console.error(error.code + " " + error.message);
-      return encodedRedirect("error", "/signup", error.message);
-    } else {
-      return encodedRedirect(
-        "success",
-        "/signup",
-        "Thanks for signing up! Please check your email for a verification link.",
-      );
-    }
-  };
-
   if ("message" in searchParams) {
     return (
       <div className="w-full flex-1 flex items-center h-screen sm:max-w-md justify-center gap-2 p-4">
@@ -67,7 +34,7 @@ export default function Signup({ searchParams }: { searchParams: Message }) {
           minLength={6}
           required
         />
-        <SubmitButton formAction={signUp} pendingText="Signing up...">
+        <SubmitButton formAction={signUpAction} pendingText="Signing up...">
           Sign up
         </SubmitButton>
         <FormMessage message={searchParams} />
